feat: inventive modifier
