--- Games/types/Mafia/roles/Village/Inventor.js
@@ -1,9 +1,0 @@
-const Role = require("../../Role");
-
-module.exports = class Inventor extends Role {
-  constructor(player, data) {
-    super("Inventor", player, data);
-    this.alignment = "Village";
-    this.cards = ["VillageCore", "WinWithVillage", "ForageItem"];
-  }
-};
