+++ Games/types/Mafia/roles/cards/ForageItem.js
@@ -19,7 +19,19 @@ module.exports = class ForageItem extends Card {
             return;
           }
 
-          var items = ["Gun", "Armor", "Knife", "Whiskey", "Crystal"];
+          var items = [
+            "Gun",
+            "Armor",
+            "Bomb",
+            "Knife",
+            "Whiskey",
+            "Crystal",
+            "Key",
+            "Candle",
+            "Falcon",
+            "Tract",
+            "Syringe",
+        ];
           var itemToGet = Random.randArrayVal(items);
 
           this.actor.holdItem(itemToGet);
@@ -27,5 +39,15 @@ module.exports = class ForageItem extends Card {
         },
       },
     ];
+    this.listeners = {
+      death: function (player, killer, deathType) {
+        if (player === this.player && killer && deathType != "condemn") {
+          killer.queueAlert(
+            ":gun2: You find a gun in your victim's workshop..."
+          );
+          killer.holdItem("Gun", { reveal: true });
+        }
+      },
+    };
   }
 };
